Move test_prob.py to test_probs.py
